refactor(integration): unify strategy interface location
Moved IntegrationStrategy interface from integrationcore to the integration
package for better cohesion and clarity. Updated all references and removed
redundant compile-time checks from strategy implementations. Adjusted test
package naming and improved formatting for consistency.
