+++ test/unit/integration/integration_test.go
@@ -1,4 +1,4 @@
-package test
+package integration_test
 
 import (
 	"math"
@@ -14,8 +14,9 @@ const tolerance = 1e-6 // Tolerance for comparing float values
 func polyFunc(x float64) float64 {
 	return x * x
 }
+
 func polyFuncIntegral(a, b float64) float64 {
-	return (math.Pow(b, 3) / 3.0) - (math.Pow(a, 3) / 3.0)
+	return (b * b * b / 3.0) - (a * a * a / 3.0)
 }
 
 // More complex function: f(x) = sin(x)
@@ -23,6 +24,7 @@ func polyFuncIntegral(a, b float64) float64 {
 func sinFunc(x float64) float64 {
 	return math.Sin(x)
 }
+
 func sinFuncIntegral(a, b float64) float64 {
 	return math.Cos(a) - math.Cos(b)
 }
@@ -32,10 +34,10 @@ func sinFuncIntegral(a, b float64) float64 {
 func expFunc(x float64) float64 {
 	return math.Exp(x)
 }
+
 func expFuncIntegral(a, b float64) float64 {
 	return math.Exp(b) - math.Exp(a)
 }
-
 
 type integrationTestCase struct {
 	name         string
@@ -86,31 +88,30 @@ func TestNewtonCotesIntegration(t *testing.T) {
 		{name: "NC1_Poly_0_1_N10", strategyName: "NewtonCotesOrder1", fn: polyFunc, a: 0, b: 1, n: 10, expected: polyFuncIntegral(0, 1), customTol: 2e-3},
 		{name: "NC1_Poly_0_1_N100", strategyName: "NewtonCotesOrder1", fn: polyFunc, a: 0, b: 1, n: 100, expected: polyFuncIntegral(0, 1), customTol: 2e-5},
 		{name: "NC1_Sin_0_Pi_N20", strategyName: "NewtonCotesOrder1", fn: sinFunc, a: 0, b: math.Pi, n: 20, expected: sinFuncIntegral(0, math.Pi), customTol: 5e-3},
-		{name: "NC1_Exp_0_1_N50", strategyName: "NewtonCotesOrder1", fn: expFunc, a: 0, b: 1, n: 50, expected: expFuncIntegral(0,1), customTol: 6e-5},
+		{name: "NC1_Exp_0_1_N50", strategyName: "NewtonCotesOrder1", fn: expFunc, a: 0, b: 1, n: 50, expected: expFuncIntegral(0, 1), customTol: 6e-5},
 		{name: "NC1_Poly_1_0_N10_ReverseBounds", strategyName: "NewtonCotesOrder1", fn: polyFunc, a: 1, b: 0, n: 10, expected: polyFuncIntegral(0, 1), customTol: 2e-3}, // Integral from 0 to 1
 		{name: "NC1_ZeroInterval", strategyName: "NewtonCotesOrder1", fn: polyFunc, a: 1, b: 1, n: 10, expected: 0},
 		{name: "NC1_Error_N0", strategyName: "NewtonCotesOrder1", fn: polyFunc, a: 0, b: 1, n: 0, expectError: true},
-
 
 		// NewtonCotesOrder2 (Simpson's 1/3)
 		{name: "NC2_Poly_0_1_N10", strategyName: "NewtonCotesOrder2", fn: polyFunc, a: 0, b: 1, n: 10, expected: polyFuncIntegral(0, 1)},
 		{name: "NC2_Poly_0_1_N100", strategyName: "NewtonCotesOrder2", fn: polyFunc, a: 0, b: 1, n: 100, expected: polyFuncIntegral(0, 1)},
 		{name: "NC2_Sin_0_Pi_N20", strategyName: "NewtonCotesOrder2", fn: sinFunc, a: 0, b: math.Pi, n: 20, expected: sinFuncIntegral(0, math.Pi), customTol: 7e-6},
-		{name: "NC2_Exp_0_1_N50", strategyName: "NewtonCotesOrder2", fn: expFunc, a: 0, b: 1, n: 50, expected: expFuncIntegral(0,1), customTol: 2e-9},
+		{name: "NC2_Exp_0_1_N50", strategyName: "NewtonCotesOrder2", fn: expFunc, a: 0, b: 1, n: 50, expected: expFuncIntegral(0, 1), customTol: 2e-9},
 		{name: "NC2_Error_N_Odd", strategyName: "NewtonCotesOrder2", fn: polyFunc, a: 0, b: 1, n: 9, expectError: true},
 
 		// NewtonCotesOrder3 (Simpson's 3/8)
 		{name: "NC3_Poly_0_1_N12", strategyName: "NewtonCotesOrder3", fn: polyFunc, a: 0, b: 1, n: 12, expected: polyFuncIntegral(0, 1)},
 		{name: "NC3_Poly_0_1_N99", strategyName: "NewtonCotesOrder3", fn: polyFunc, a: 0, b: 1, n: 99, expected: polyFuncIntegral(0, 1)},
 		{name: "NC3_Sin_0_Pi_N21", strategyName: "NewtonCotesOrder3", fn: sinFunc, a: 0, b: math.Pi, n: 21, expected: sinFuncIntegral(0, math.Pi), customTol: 1.3e-5},
-		{name: "NC3_Exp_0_1_N51", strategyName: "NewtonCotesOrder3", fn: expFunc, a: 0, b: 1, n: 51, expected: expFuncIntegral(0,1), customTol: 4e-9},
+		{name: "NC3_Exp_0_1_N51", strategyName: "NewtonCotesOrder3", fn: expFunc, a: 0, b: 1, n: 51, expected: expFuncIntegral(0, 1), customTol: 4e-9},
 		{name: "NC3_Error_N_NotMult3", strategyName: "NewtonCotesOrder3", fn: polyFunc, a: 0, b: 1, n: 10, expectError: true},
 
 		// NewtonCotesOrder4 (Boole's)
 		{name: "NC4_Poly_0_1_N12", strategyName: "NewtonCotesOrder4", fn: polyFunc, a: 0, b: 1, n: 12, expected: polyFuncIntegral(0, 1), customTol: 1e-9},
 		{name: "NC4_Poly_0_1_N100", strategyName: "NewtonCotesOrder4", fn: polyFunc, a: 0, b: 1, n: 100, expected: polyFuncIntegral(0, 1), customTol: 1e-10},
 		{name: "NC4_Sin_0_Pi_N20", strategyName: "NewtonCotesOrder4", fn: sinFunc, a: 0, b: math.Pi, n: 20, expected: sinFuncIntegral(0, math.Pi), customTol: 7e-8},
-		{name: "NC4_Exp_0_1_N52", strategyName: "NewtonCotesOrder4", fn: expFunc, a: 0, b: 1, n: 52, expected: expFuncIntegral(0,1), customTol: 1e-11},
+		{name: "NC4_Exp_0_1_N52", strategyName: "NewtonCotesOrder4", fn: expFunc, a: 0, b: 1, n: 52, expected: expFuncIntegral(0, 1), customTol: 1e-11},
 		{name: "NC4_Error_N_NotMult4", strategyName: "NewtonCotesOrder4", fn: polyFunc, a: 0, b: 1, n: 10, expectError: true},
 	}
 
@@ -129,22 +130,22 @@ func TestGaussLegendreIntegration(t *testing.T) {
 		// GaussLegendreOrder1
 		{name: "GL1_Poly_0_1", strategyName: "GaussLegendreOrder1", fn: polyFunc, a: 0, b: 1, n: 1, expected: polyFuncIntegral(0, 1), customTol: 1e-1}, // Low accuracy expected
 		{name: "GL1_Sin_0_Pi", strategyName: "GaussLegendreOrder1", fn: sinFunc, a: 0, b: math.Pi, n: 1, expected: sinFuncIntegral(0, math.Pi), customTol: 1.2},
-        {name: "GL1_Exp_0_1", strategyName: "GaussLegendreOrder1", fn: expFunc, a:0, b:1, n:1, expected: expFuncIntegral(0,1), customTol: 7e-2},
+		{name: "GL1_Exp_0_1", strategyName: "GaussLegendreOrder1", fn: expFunc, a: 0, b: 1, n: 1, expected: expFuncIntegral(0, 1), customTol: 7e-2},
 
 		// GaussLegendreOrder2
 		{name: "GL2_Poly_0_1", strategyName: "GaussLegendreOrder2", fn: polyFunc, a: 0, b: 1, n: 1, expected: polyFuncIntegral(0, 1)}, // Exact for cubics
 		{name: "GL2_Sin_0_Pi", strategyName: "GaussLegendreOrder2", fn: sinFunc, a: 0, b: math.Pi, n: 1, expected: sinFuncIntegral(0, math.Pi), customTol: 6.5e-2},
-        {name: "GL2_Exp_0_1", strategyName: "GaussLegendreOrder2", fn: expFunc, a:0, b:1, n:1, expected: expFuncIntegral(0,1), customTol: 4e-4},
+		{name: "GL2_Exp_0_1", strategyName: "GaussLegendreOrder2", fn: expFunc, a: 0, b: 1, n: 1, expected: expFuncIntegral(0, 1), customTol: 4e-4},
 
 		// GaussLegendreOrder3
 		{name: "GL3_Poly_0_1", strategyName: "GaussLegendreOrder3", fn: polyFunc, a: 0, b: 1, n: 1, expected: polyFuncIntegral(0, 1)}, // Exact for quintics
 		{name: "GL3_Sin_0_Pi", strategyName: "GaussLegendreOrder3", fn: sinFunc, a: 0, b: math.Pi, n: 1, expected: sinFuncIntegral(0, math.Pi), customTol: 1.4e-3},
-        {name: "GL3_Exp_0_1", strategyName: "GaussLegendreOrder3", fn: expFunc, a:0, b:1, n:1, expected: expFuncIntegral(0,1), customTol: 1e-6},
+		{name: "GL3_Exp_0_1", strategyName: "GaussLegendreOrder3", fn: expFunc, a: 0, b: 1, n: 1, expected: expFuncIntegral(0, 1), customTol: 1e-6},
 
 		// GaussLegendreOrder4
 		{name: "GL4_Poly_0_1", strategyName: "GaussLegendreOrder4", fn: polyFunc, a: 0, b: 1, n: 1, expected: polyFuncIntegral(0, 1)}, // Exact for 7th degree polynomials
 		{name: "GL4_Sin_0_Pi", strategyName: "GaussLegendreOrder4", fn: sinFunc, a: 0, b: math.Pi, n: 1, expected: sinFuncIntegral(0, math.Pi), customTol: 1.6e-5},
-        {name: "GL4_Exp_0_1", strategyName: "GaussLegendreOrder4", fn: expFunc, a:0, b:1, n:1, expected: expFuncIntegral(0,1), customTol: 1e-9},
+		{name: "GL4_Exp_0_1", strategyName: "GaussLegendreOrder4", fn: expFunc, a: 0, b: 1, n: 1, expected: expFuncIntegral(0, 1), customTol: 1e-9},
 		{name: "GL4_Poly_ZeroInterval", strategyName: "GaussLegendreOrder4", fn: polyFunc, a: 1, b: 1, n: 1, expected: 0},
 	}
